refactor(admission): split hardware admission modules
